+++ webapp/login.go
@@ -12,11 +12,9 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/google/go-github/v28/github"
 	"github.com/gorilla/securecookie"
 	"github.com/web-platform-tests/wpt.fyi/shared"
 	"golang.org/x/oauth2"
-	ghOAuth "golang.org/x/oauth2/github"
 	"google.golang.org/appengine"
 )
 
@@ -31,81 +29,6 @@ type User struct {
 	GithuhEmail  string
 }
 
-type githubOAuthImp struct {
-	ctx         context.Context
-	ds          shared.Datastore
-	conf        *oauth2.Config
-	accessToken *string
-}
-
-func (g *githubOAuthImp) Datastore() shared.Datastore {
-	return g.ds
-}
-
-func (g *githubOAuthImp) Context() context.Context {
-	return g.ctx
-}
-
-func (g *githubOAuthImp) GetAccessToken() *string {
-	return g.accessToken
-}
-
-func (g *githubOAuthImp) SetRedirectURL(url string) {
-	g.conf.RedirectURL = url
-}
-
-func (g *githubOAuthImp) GetAuthCodeURL(state string, opts ...oauth2.AuthCodeOption) string {
-	return g.conf.AuthCodeURL(state, opts...)
-}
-
-func (g *githubOAuthImp) GetNewClient(oauthToken string) (*github.Client, error) {
-	token, err := g.conf.Exchange(g.ctx, oauthToken)
-	if err != nil {
-		return nil, err
-	}
-	g.accessToken = &token.AccessToken
-
-	oauthClient := oauth2.NewClient(g.ctx, oauth2.StaticTokenSource(token))
-	client := github.NewClient(oauthClient)
-
-	return client, nil
-}
-
-func (g *githubOAuthImp) GetGitHubUser(client *github.Client) (*github.User, error) {
-	ghUser, _, err := client.Users.Get(g.ctx, "")
-	if err != nil {
-		return nil, err
-	}
-
-	return ghUser, nil
-}
-
-func newGitHubOAuth(ctx context.Context) (shared.GitHubOAuth, error) {
-	store := shared.NewAppEngineDatastore(ctx, false)
-	log := shared.GetLogger(ctx)
-	clientID, err := shared.GetSecret(store, "github-oauth-client-id")
-	if err != nil {
-		log.Errorf("Failed to get github-oauth-client-id secret: %s", err.Error())
-		return &githubOAuthImp{}, err
-	}
-
-	secret, err := shared.GetSecret(store, "github-oauth-client-secret")
-	if err != nil {
-		log.Errorf("Failed to get github-oauth-client-secret: %s", err.Error())
-		return &githubOAuthImp{}, err
-	}
-
-	oauth := &oauth2.Config{
-		ClientID:     clientID,
-		ClientSecret: secret,
-		// (no scope) - see https://developer.github.com/apps/building-oauth-apps/understanding-scopes-for-oauth-apps/#available-scopes
-		Scopes:   []string{},
-		Endpoint: ghOAuth.Endpoint,
-	}
-
-	return &githubOAuthImp{ctx: ctx, conf: oauth, ds: store}, nil
-}
-
 func loginHandler(w http.ResponseWriter, r *http.Request) {
 	ctx := shared.NewAppEngineContext(r)
 	aeAPI := shared.NewAppEngineAPI(ctx)
@@ -114,7 +37,7 @@ func loginHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	githubOauthImp, err := newGitHubOAuth(ctx)
+	githubOauthImp, err := shared.NewGitHubOAuth(ctx)
 	if err != nil {
 		http.Error(w, "Error creating githuboauthImp", http.StatusInternalServerError)
 		return
@@ -164,7 +87,7 @@ func handleLogin(g shared.GitHubOAuth, w http.ResponseWriter, r *http.Request) {
 
 func oauthHandler(w http.ResponseWriter, r *http.Request) {
 	ctx := shared.NewAppEngineContext(r)
-	githuboauthImp, err := newGitHubOAuth(ctx)
+	githuboauthImp, err := shared.NewGitHubOAuth(ctx)
 	if err != nil {
 		http.Error(w, "Error creating githuboauthImp", http.StatusInternalServerError)
 		return
